+++ bu-blocks.php
@@ -7,7 +7,7 @@
  * Author URI:      http://www.bu.edu/interactive-design/
  * Text Domain:     bu-blocks
  * Domain Path:     /languages
- * Version:         0.3.3
+ * Version:         0.3.4
  *
  * @package         BU_Blocks
  */
@@ -20,7 +20,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 
 // Defines the plugin version.
-define( 'BU_BLOCKS_VERSION', '0.3.3' );
+define( 'BU_BLOCKS_VERSION', '0.3.4' );
 
 /**
  * Displays admin notice and prevents activation.